FIRM panel and other fields update (not final)
Attempts to update the firm panel and flood information for CSLT parcels.
This is a very slow script because of the multiple MakeFeatureLayer. This might be useful later to handle populating fields after the parcel core is updated. Typically only 6-10 rows will need attention. 

A top-down approach to updating every parcel will be similar to the plan area logic