Create contourAnalysis
Script that seeks to decrease the size of a large 1ft contour file. Very short contours need to be removed, but we dont want to lose any data. The script steps through contours in 1m intervals, and evaluates what percentage of the total file will be removed. 
In this script you set the desired reduction ratio and starting length.
Finally, matplotlib.pylab is used to create a graph of the analysis so you can see the point of diminishing returns.