Basic data export script
The only user input required is a boundary ploygon and folder name.
Script selects all stormwater FC's and exports them to shapefiles. Names are the same as the original FCs.
Final part of script attempts to analyze the results and identify any FCs that didn't make it. The first test is a simple comparison of counts in the starting list and list from a ListFeatureClasses() in the output directory.

Expansion plans:
add option for gdb export, including putting the FCs in a feature dataset
add ability to include a .lyr file to make the receivers life easier
create another version of the script to run with arcmap open
add ability for user to add FCs from other directories